use longs

+++ src/main/java/org/piki/App.java
@@ -8,12 +8,12 @@ public class App {
         }
     }
 
-    public static int fibonacci(int n) {
+    public static long fibonacci(int n) {
         if (n <= 2) return 1;
 
-        int a = 1, b = 1;
+        long a = 1, b = 1;
         for (int i=2; i<n; i++) {
-            int t = a;
+            long t = a;
             a = b;
             b = b + t;
         }
